+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## unreleased
+### Fixed
+* Fixed an issue where exporting would fail if an image texture node did not have the correct label.
+
 ## 0.13.0 - 2024-08-16
 ### Added
 * Added an option to replaces textures with assigned images from current scene for wimdo export.